Create project task request to contracts
