+++ src/core/Codend.Contracts/Requests/ProjectTasks/Create/CreateBugfixProjectTaskRequest.cs
@@ -1,0 +1,16 @@
+namespace Codend.Contracts.Requests.ProjectTasks.Create;
+
+/// <inheritdoc />
+public record CreateBugfixProjectTaskRequest
+(
+    string Name,
+    string Priority,
+    Guid StatusId,
+    Guid ProjectId,
+    string? Description,
+    EstimatedTimeRequest? EstimatedTime,
+    DateTime? DueDate,
+    uint? StoryPoints,
+    Guid? AssigneeId,
+    Guid? StoryId
+) : ICreateBaseProjectTaskRequest;
