--- src/core/Codend.Contracts/Requests/ProjectTasks/Create/ICreateBugfixProjectTaskRequest.cs
@@ -1,8 +1,0 @@
-namespace Codend.Contracts.Requests.ProjectTasks.Create;
-
-/// <summary>
-/// Request used for creating BugfixProjectTask 
-/// </summary>
-public interface ICreateBugfixProjectTaskRequest : ICreateBaseProjectTaskRequest
-{
-}
